feat(gb-9678): unified interface types to avoid complexity later on
