+++ crates/llm/src/server/handler.rs
@@ -1,7 +1,10 @@
 //! LLM handler that conditionally applies metrics and tracing
 
 use crate::{
-    messages::openai::{ChatCompletionRequest, ChatCompletionResponse, ModelsResponse},
+    messages::{
+        openai::ModelsResponse,
+        unified::{UnifiedRequest, UnifiedResponse},
+    },
     provider::ChatCompletionStream,
     request::RequestContext,
     server::{LlmServer, LlmService, metrics::LlmServerWithMetrics, tracing::LlmServerWithTracing},
@@ -31,31 +34,31 @@ impl LlmHandler {
         }
     }
 
-    /// Process a chat completion request.
+    /// Process a unified chat completion request (protocol-agnostic).
     pub async fn completions(
         &self,
-        request: ChatCompletionRequest,
+        request: UnifiedRequest,
         context: &RequestContext,
-    ) -> crate::Result<ChatCompletionResponse> {
+    ) -> crate::Result<UnifiedResponse> {
         match self {
             LlmHandler::WithMetricsAndTracing(server) => server.completions(request, context).await,
             LlmHandler::WithMetrics(server) => server.completions(request, context).await,
             LlmHandler::WithTracing(server) => server.completions(request, context).await,
-            LlmHandler::Direct(server) => server.completions(request, context).await,
+            LlmHandler::Direct(server) => server.unified_completions(request, context).await,
         }
     }
 
-    /// Process a streaming chat completion request.
+    /// Process a unified streaming chat completion request (protocol-agnostic).
     pub async fn completions_stream(
         &self,
-        request: ChatCompletionRequest,
+        request: UnifiedRequest,
         context: &RequestContext,
     ) -> crate::Result<ChatCompletionStream> {
         match self {
             LlmHandler::WithMetricsAndTracing(server) => server.completions_stream(request, context).await,
             LlmHandler::WithMetrics(server) => server.completions_stream(request, context).await,
             LlmHandler::WithTracing(server) => server.completions_stream(request, context).await,
-            LlmHandler::Direct(server) => server.completions_stream(request, context).await,
+            LlmHandler::Direct(server) => server.unified_completions_stream(request, context).await,
         }
     }
 }
